Not bad results with Listening mode. Idle 262ms*16=4.192s Rx 262ms*32=8.382s RxTimeout2= 255*16*Tbit=71ms after RSSI RSSIThreshold = 160 = -80dB TxPower = 20dB Tx cunsumtion is almost minimal. 0.5mA in Sleep + 1 packet in 4 seconds Rx consumption is minimal too. 0.5mA in Idle + shortest Rx period 71ms of active receiving

+++ firmware/lcd.c
@@ -39,7 +39,7 @@ void LCD_TransmitDot(char cData, char isdot)
     char tData = 0;
     if (cData < sizeof(CHARTABLE)/sizeof(CHARTABLE[0]))
     {
-        tData = pgm_read_byte(&CHARTABLE[cData]);
+        tData = pgm_read_byte(&CHARTABLE[(int)cData]);
     }
     if( isdot & LCD_DOT){
         SPI_SetData(0x80);
@@ -59,7 +59,7 @@ void LCD_Transmit(char cData)
     char tData = 0;
     if (cData < sizeof(CHARTABLE)/sizeof(CHARTABLE[0]))
     {
-        tData = pgm_read_byte(&CHARTABLE[cData]);
+        tData = pgm_read_byte(&CHARTABLE[(int)cData]);
     }
     
     SPI_SetData(0x00);
